Add files via upload
# Changelog

## v1.4.0

### 数据导入导出

- 新增 Tab **「📦 数据导入导出」**，所有操作在界面内完成，无需外部工具。
- **导出**：支持 Markdown、CSV（兼容 Excel）、SQL 三种格式。冰箱表含 ID 列，物品表含冰箱ID 列。导出文件含日期时间和单位说明注释。
- **导入**：支持同三种格式导入。冰箱按名称跳过重复，物品全部直接新增。CSV 注释行自动跳过，Markdown 支持 HTML 注释和 ID/名称两种冰箱匹配方式。
- **空白模板**：Markdown 和 CSV 导入区可下载空白模板，保留现有冰箱数据（含 ID）供参考，物品表为空。

### 库存浏览增强

- **排序**：新增排序下拉框，7 种方式—到期时间（默认）、最近/最早添加、名称 A-Z/Z-A、数量多→少/少→多。切换排序自动回第 1 页。
- **筛选**：新增存储类型下拉（冷冻/冷藏），与冰箱、分类、搜索自由组合。
- **快速取用**：物品卡片新增绿色 −1 按钮（`api/consume.php`），点击减 1，数量归零自动删除。
- **确认弹窗**：取用和导入操作使用页面内自定义弹窗替代浏览器 `confirm()`，兼容手机端。

### 物品转移优化

- 新增分类筛选下拉和名称搜索框，物品多时可快速定位。
- 执行转移按钮移到物品列表上方，物品再多也始终可见。
- 手机竖屏适配。

### 数量精度

- `quantity` 字段从 INT 改为 DECIMAL(10,2)，支持小数重量（如 1.5 斤、0.8 把）。
- 全链路 `parseInt`/`(int)` 改为 `parseFloat`/`(float)`，输入框 `step="0.01"`。
- 响应层数量格式化：整数不显示 .00，小数保留实际精度。
- 新增 `upgrade.php` 浏览器一键升级，`migrate_v3.sql` 供手动迁移。

### Bug 修复

- CSV 导入注释行被引号包裹导致未能识别 → 改为 `str_getcsv` 解析后再匹配。
- Markdown 导入 HTML 注释中断表头解析 → 移除 `$inTable` 状态机，简化为直接跳过非 `|` 行。
- 到期时间排序不生效 → SQL ORDER BY 的 `ELSE NULL` 与响应层兜底日期不一致，改为 `ELSE DATE_ADD(...storage_type...)`。

---

## v1.3.0 (2026-05-23)

### Web 安装向导

- 新增 `install.php` 一键部署向导，浏览器填写数据库信息即可完成安装。
- 安装完成后生成 `install.lock`，再次访问拒绝运行。
- 检测已有数据时提供保留/覆盖/取消三种处理方式。
- `init.sql` 分类预置改用 `INSERT IGNORE`，保证幂等。

### 过期推算修复

- 有生产日期但无保质期的物品，统一按存储类型兜底推算（冷藏 +5 天 / 冷冻 +90 天）。

---

+++ api/transfer.php
@@ -26,7 +26,7 @@ switch ($method) {
             jsonResponse(['error' => '来源和目标冰箱不能相同'], 400);
         }
 
-        $quantity = (int)($data['quantity'] ?? 1);
+        $quantity = (float)($data['quantity'] ?? 1);
 
         // 获取物品当前信息
         $stmt = $db->prepare('SELECT * FROM items WHERE id = ?');
